// core/types/generated/contentful.d.ts
export interface IAuthorFields {
  /** Author Name */
  authorName: string;

  /** Author Image */
  authorImage?: Asset | undefined;
}

export interface IAuthor extends Entry<IAuthorFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "author";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ICarouselProductFields {
  /** Internal Name */
  internalName: string;

  /** Carousel Title */
  carouselTitle: string;

  /** Products */
  products: IProductFinishedGoods[];
}

export interface ICarouselProduct extends Entry<ICarouselProductFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "carouselProduct";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ICarouselRecipeFields {
  /** Internal Name */
  internalName: string;

  /** Carousel Title */
  carouselTitle: string;

  /** Recipes */
  recipes: IRecipe[];
}

export interface ICarouselRecipe extends Entry<ICarouselRecipeFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "carouselRecipe";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ICtaFields {
  /** Text */
  text: string;

  /** Internal Reference */
  internalReference?:
    | IProductFinishedGoods
    | IProductPartsAndAccessories
    | IPageStandard
    | undefined;

  /** External Link */
  externalLink?: string | undefined;
}

export interface ICta extends Entry<ICtaFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "cta";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IFeatureVideoBannerFields {
  /** Internal Name */
  internalName: string;

  /** Section Title */
  sectionTitle?: string | undefined;

  /** Sub Title */
  subTitle?: string | undefined;

  /** Descriptive Body Copy */
  descriptiveBodyCopy?: string | undefined;

  /** Video */
  video: Asset;

  /** Registration Cookie Message */
  registrationCookieMessage?: string | undefined;
}

export interface IFeatureVideoBanner extends Entry<IFeatureVideoBannerFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "featureVideoBanner";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IIngredientsListFields {
  /** Ingredients List Name (Internal) */
  ingredientsListName: string;

  /** Section Title */
  sectionTitle?: string | undefined;

  /** List of Ingredients */
  listOfIngredients?: string[] | undefined;
}

export interface IIngredientsList extends Entry<IIngredientsListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "ingredientsList";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IInspirationCardFields {
  /** Title */
  title: string;

  /** Content Reference */
  contentReference: ITutorial | IRecipe;
}

export interface IInspirationCard extends Entry<IInspirationCardFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "inspirationCard";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IPageStandardFields {
  /** Page Name */
  pageName: string;

  /** Meta Title (SEO) */
  metaTitleSeo: string;

  /** Meta Description (SEO) */
  metaDescription?: string | undefined;

  /** Meta Keywords (SEO) */
  metaKeywordsSeo?: string | undefined;

  /** Page Slug */
  pageSlug: string;

  /** Optional Page Description */
  optionalPageDescription?: Document | undefined;

  /** Page Content Entries */
  pageContent?: Entry<{ [fieldId: string]: unknown }>[] | undefined;
}

export interface IPageStandard extends Entry<IPageStandardFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "pageStandard";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IProductFinishedGoodsFields {
  /** Product Name */
  productName: string;

  /** BC Product Reference */
  bcProductReference: string;

  /** Page Slug */
  pageSlug?: string | undefined;

  /** Short Description */
  shortDescription?: string | undefined;

  /** Default Price */
  defaultPrice: string;

  /** Sale Price */
  salePrice?: string | undefined;

  /** Sale Message */
  couponCodesalesDates?: string | undefined;

  /** Details */
  details?: Document | undefined;

  /** FAQs */
  faqs?: IFaq[] | undefined;

  /** Featured Image */
  featuredImage?: Asset | undefined;

  /** Additional Images */
  additionalImages?: Asset[] | undefined;

  /** Parts + Accessories */
  partsAccessories?: Entry<{ [fieldId: string]: unknown }>[] | undefined;

  /** Product Carousel */
  productCarousel?: ICarouselProduct[] | undefined;

  /** Meta Title */
  metaTitle?: string | undefined;

  /** Meta Description */
  metaDescription?: string | undefined;

  /** Product Line */
  productLine?: string[] | undefined;

  /** Parent Category */
  parentCategory?: string[] | undefined;

  /** Sub-Category */
  subCategory?: string[] | undefined;

  /** Web Sub Category */
  webSubCategory?: string | undefined;

  /** Product Formulation Information */
  productFormulationInformation?: Record<string, any> | undefined;

  /** Feature */
  feature?: string[] | undefined;

  /** Finish */
  finish?: string[] | undefined;

  /** Size */
  size?: string | undefined;

  /** Support Documentation */
  supportDocumentation?: ISupportDocument[] | undefined;

  /** Factory Recertified Product */
  factoryRecertifiedProduct?: boolean | undefined;

  /** Model Number */
  modelNumber: string;

  /** Wattage */
  wattage?: string | undefined;

  /** Warranty */
  warranty?: string | undefined;

  /** Spotlight Video */
  spotlightVideo?: IFeatureVideoBanner | undefined;

  /** Out of Box net_weight */
  outOfBoxNetWeight?: string | undefined;

  /** Out of Box depth */
  outOfBoxDepth?: string | undefined;

  /** Out of Box width */
  outOfBoxWidth?: string | undefined;

  /** Out of Box height */
  outOfBoxHeight?: string | undefined;

  /** Out of Box Size UOM */
  outOfBoxSizeUom?: string | undefined;

  /** Out of Box Weight UOM */
  outOfBoxWeightUom?: string | undefined;

  /** Archive */
  archive?: boolean | undefined;

  /** Product Badge */
  productBadge?: string | undefined;

  /** isShipsFree */
  isShipsFree?: boolean | undefined;

  /** Inventory Quantity */
  inventoryQuantity?: number | undefined;

  /** Recipes */
  recipes?: ICarouselRecipe | undefined;

  /** Page Content Entries */
  pageContentEntries?:
    | (IBlockProductFeatures | IBlockProductFeaturesAccordion)[]
    | undefined;
}

export interface IProductFinishedGoods
  extends Entry<IProductFinishedGoodsFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "productFinishedGoods";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IProductPartsAndAccessories
  extends Entry<IProductPartsAndAccessoriesFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "productPartsAndAccessories";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IRecipeFields {
  /** Recipe Name */
  recipeName: string;

  /** Meta Title */
  metaTitle: string;

  /** Meta Description */
  metaDescription: string;

  /** Page Slug */
  pageSlug: string;

  /** Short Description */
  shortDescription?: string | undefined;

  /** Meal Type (Category) */
  mealTypeCategory?:
    | (
        | "Appetizers"
        | "Beverages"
        | "Breakfast Foods"
        | "Desserts"
        | "Dressings, Sauces, Marinades"
        | "Main Course"
        | "Salads"
        | "Sandwiches"
        | "Side Dishes"
        | "Snack"
        | "Soups, Stews and Chili"
        | "Dinner"
      )[]
    | undefined;

  /** Occasion (Category) */
  occasionCategory?:
    | (
        | "4th of July"
        | "Cinco de Mayo"
        | "Easter"
        | "Halloween"
        | "Labor Day"
        | "Memorial Day"
        | "New Year's"
        | "Thanksgiving"
        | "Christmas"
      )[]
    | undefined;

  /** Ingredients (Category) */
  ingredientsCategory?:
    | (
        | "Beef"
        | "Chicken"
        | "Eggs"
        | "Fruit"
        | "Lamb"
        | "Pasta, Rice & Beans"
        | "Pork"
        | "Sausage"
        | "Seafood"
        | "Turkey"
        | "Vegetables"
        | "Game Meats"
      )[]
    | undefined;

  /** Appliance Type (Category) */
  applianceTypeCategory?:
    | (
        | "Burger Presses"
        | "Butcher Saws & Knives"
        | "Canning"
        | "Deep Fryers"
        | "Dehydrators"
        | "Freezer Paper & Bags"
        | "French Fry Cutters"
        | "Grills"
        | "Jerky Guns"
        | "Meat Grinders"
        | "Meat Injector"
        | "Meat Slicers"
        | "Roaster Ovens"
        | "Sausage Casings"
        | "Sausage Stuffers"
        | "Seasonings"
        | "Smoker/Slow Cookers"
        | "Strainers"
        | "Vacuum Sealer Bags"
        | "Vacuum Sealers"
      )[]
    | undefined;

  /** Author */
  author?: IAuthor | undefined;

  /** Cook Time */
  cookTime?: string | undefined;

  /** Number of Ingredients */
  numberOfIngredients?: string | undefined;

  /** Number of Servings */
  numberOfServings?: string | undefined;

  /** Ingredients List(s) */
  ingredientsLists?: IIngredientsList[] | undefined;

  /** Recipe Directions */
  recipeDirections?: Document | undefined;

  /** Test Kitchen Tips */
  testKitchenTips?: Document | undefined;

  /** Featured Image */
  featuredImage: Asset;

  /** Additional Images */
  additionalImages?: Asset[] | undefined;

  /** Video Feature */
  videoFeature?: IFeatureVideoBanner | undefined;

  /** Product Carousel */
  productCarousel?: ICarouselProduct | undefined;

  /** Recipe Carousel */
  recipeCarousel?: ICarouselRecipe | undefined;
}

export interface IRecipe extends Entry<IRecipeFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "recipe";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ITutorialFields {
  /** Title */
  title: string;
}

export interface ITutorial extends Entry<ITutorialFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "tutorial";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}